Fix some linter issues

+++ linux/hci/signal.go
@@ -29,10 +29,10 @@ func (c *Conn) Signal(req Signal, rsp Signal) error {
 	data := req.Marshal()
 	buf := bytes.NewBuffer(make([]byte, 0))
 	binary.Write(buf, binary.LittleEndian, uint16(4+len(data)))
-	binary.Write(buf, binary.LittleEndian, uint16(cidLESignal))
+	binary.Write(buf, binary.LittleEndian, cidLESignal)
 
 	binary.Write(buf, binary.LittleEndian, uint8(req.Code()))
-	binary.Write(buf, binary.LittleEndian, uint8(c.sigID))
+	binary.Write(buf, binary.LittleEndian, c.sigID)
 	binary.Write(buf, binary.LittleEndian, uint16(len(data)))
 	binary.Write(buf, binary.LittleEndian, data)
 
@@ -66,9 +66,9 @@ func (c *Conn) sendResponse(code uint8, id uint8, r Signal) (int, error) {
 	data := r.Marshal()
 	buf := bytes.NewBuffer(make([]byte, 0))
 	binary.Write(buf, binary.LittleEndian, uint16(4+len(data)))
-	binary.Write(buf, binary.LittleEndian, uint16(cidLESignal))
-	binary.Write(buf, binary.LittleEndian, uint8(code))
-	binary.Write(buf, binary.LittleEndian, uint8(id))
+	binary.Write(buf, binary.LittleEndian, cidLESignal)
+	binary.Write(buf, binary.LittleEndian, code)
+	binary.Write(buf, binary.LittleEndian, id)
 	binary.Write(buf, binary.LittleEndian, uint16(len(data)))
 	if err := binary.Write(buf, binary.LittleEndian, data); err != nil {
 		return 0, err
